Refactor SQL Server Command handlers into own class
