+++ src/GUI/SqlCe35Toolbox/VSToolbox.csproj
@@ -182,6 +182,7 @@
     <Compile Include="Commands\ColumnMenuCommands.cs" />
     <Compile Include="Commands\ColumnMenuCommandsHandler.cs" />
     <Compile Include="Commands\CommandHandlerBase.cs" />
+    <Compile Include="Commands\SqlServerDatabaseMenuCommandsHandler.cs" />
     <Compile Include="Commands\TriggerMenuCommandsHandler.cs" />
     <Compile Include="Commands\ViewMenuCommandsHandler.cs" />
     <Compile Include="Commands\DatabasesMenuCommandsHandler.cs" />
